Model is now split in 4 files: Product.cs and ProductOption.cs are model files matching the structure presented in README (and the database). These has been self generated with EF Products.cs and ProductOption.cs are model files for returning a list of products/productionoptions in a Items array (as presented in README)
